tox_callback_name_change()
toxcore eac0d435f35db468435fa596a3b00b593d6bf3e9

+++ src/toxic.h
@@ -102,7 +102,7 @@ void on_request(Tox *m, const uint8_t *public_key, const uint8_t *data, uint16_t
 void on_connectionchange(Tox *m, uint32_t friendnumber, TOX_CONNECTION connection_status, void *userdata);
 void on_message(Tox *m, int32_t friendnumber, const uint8_t *string, uint16_t length, void *userdata);
 void on_action(Tox *m, int32_t friendnumber, const uint8_t *string, uint16_t length, void *userdata);
-void on_nickchange(Tox *m, int32_t friendnumber, const uint8_t *string, uint16_t length, void *userdata);
+void on_nickchange(Tox *m, uint32_t friendnumber, const uint8_t *string, size_t length, void *userdata);
 void on_statuschange(Tox *m, int32_t friendnumber, uint8_t status, void *userdata);
 void on_statusmessagechange(Tox *m, int32_t friendnumber, const uint8_t *string, uint16_t length, void *userdata);
 void on_friendadded(Tox *m, int32_t friendnumber, bool sort);
